feat: Implement theme selector
