+++ components/Modal.js
@@ -1,9 +1,9 @@
 import { Fragment } from "react";
 import { Dialog } from "@headlessui/react";
 
-export default function Modal({ open, setOpen, children }) {
+export default function Modal({ open, setOpen, children, settings }) {
   return (
-      <Dialog as="div" className="modal theme--green" open={open} onClose={setOpen}>
+      <Dialog as="div" className={`modal theme--${settings.theme}`} open={open} onClose={setOpen}>
         <Dialog.Panel className="modal__content">
           <div className="modal__body">
             {children}
